scheme: Replace UUID with Simple Ids

+++ library/IvozProvider/Mapper/Sql/DbTable/BrandURLs.php
@@ -54,7 +54,7 @@ class BrandURLs extends TableAbstract
 	    'TABLE_NAME' => 'BrandURLs',
 	    'COLUMN_NAME' => 'id',
 	    'COLUMN_POSITION' => 1,
-	    'DATA_TYPE' => 'mediumint',
+	    'DATA_TYPE' => 'int',
 	    'DEFAULT' => NULL,
 	    'NULLABLE' => false,
 	    'LENGTH' => NULL,
