+++ library/IvozProvider/Mapper/Sql/DbTable/TargetGroups.php
@@ -57,7 +57,7 @@ class TargetGroups extends TableAbstract
 	    'TABLE_NAME' => 'TargetGroups',
 	    'COLUMN_NAME' => 'id',
 	    'COLUMN_POSITION' => 1,
-	    'DATA_TYPE' => 'mediumint',
+	    'DATA_TYPE' => 'int',
 	    'DEFAULT' => NULL,
 	    'NULLABLE' => false,
 	    'LENGTH' => NULL,
